pyproject.toml
[project]
name = "quinn-voice-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "av==17.0.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.5",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.5",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "eval-type-backport==0.3.1",
    "filelock==3.25.2",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "google>=3.0.0",
    "google-auth==2.49.1",
    "google-genai==1.67.0",
    "googleapis-common-protos==1.73.0",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.36.2",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "livekit==1.1.2",
    "livekit-agents[cartesia,deepgram,silero,turn-detector]==1.4.5",
    "livekit-api==1.1.0",
    "livekit-blingfire==1.1.0",
    "livekit-plugins-cartesia==1.4.5",
    "livekit-plugins-deepgram==1.4.5",
    "livekit-plugins-silero==1.4.5",
    "livekit-plugins-turn-detector==1.4.5",
    "livekit-protocol==1.1.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "nest-asyncio==1.6.0",
    "numpy==2.4.3",
    "onnxruntime==1.24.3",
    "openai==2.28.0",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-exporter-otlp-proto-http==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "packaging==26.0",
    "prometheus-client==0.24.1",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "psycopg2>=2.9.11",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyjwt==2.12.1",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
    "regex==2026.2.28",
    "requests==2.32.5",
    "rich==14.3.3",
    "safetensors==0.7.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sounddevice==0.5.5",
    "sqlalchemy>=2.0.48",
    "sympy==1.14.0",
    "tenacity==9.1.4",
    "tokenizers==0.22.2",
    "tqdm==4.67.3",
    "transformers==4.57.1",
    "typer==0.24.1",
    "types-protobuf==6.32.1.20260221",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "watchfiles==1.1.1",
    "websockets==16.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
]
